requirements.txt
numpy
python-binance
python-dotenv
requests
